Add files via upload
C# Code to build an application which Silently Copies File From USB Drive to a particular folder on internal Drive And Further can be modified for Network drive.
This C# code is for copying multiple extensions file

I had used Stackoverflow.com to build this code. and in the comment section, you can find links to that particular code which will help in better understanding of code.